requirements.txt
python-telegram-bot==22
python-dotenv==1.1.0
supabase==2.15.0
asyncio==3.4.3 